+++ outbox_store.go
@@ -8,7 +8,7 @@ import (
 	"fmt"
 	"time"
 
-	json "github.com/gtkit/json"
+	json "github.com/gtkit/json/v2"
 	"gorm.io/gorm"
 )